refactor(reviewer-bot): extract comment routing
